connect many more packets, properly close dirs & access maps
Use find instead of [] to check if maps have elements

filter DNS query packets

+++ src/application.h
@@ -6,14 +6,14 @@
 #include <ctime>
 
 struct application {
-    int id;            /* database application id */
-    pid_t pid;         /* pid directory for application */
-    char *name;        /* application name (pruned process cmdline) */
-    int pkt_rx;        /* packets received in bytes */
-    int pkt_tx;        /* packets transmitted in bytes */
-    int pkt_tcp;       /* number of tcp packets */
-    int pkt_udp;       /* number of udp packets */
-    time_t start_time; /* timestamp for when application detected */
+    int id;                    /* database application id */
+    pid_t pid;                 /* pid directory for application */
+    char *name;                /* application name (pruned process cmdline) */
+    unsigned long long pkt_rx; /* packets received in bytes */
+    unsigned long long pkt_tx; /* packets transmitted in bytes */
+    int pkt_tcp;               /* number of tcp packets */
+    int pkt_udp;               /* number of udp packets */
+    time_t start_time;         /* timestamp for when application detected */
 };
 
 int get_application_name(struct application *app, int pid);
